intelligent_job_finder/backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Caching
redis==5.0.1
aioredis==2.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Background Tasks
celery==5.3.4
flower==2.0.1

# AI & ML Frameworks
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
langgraph==0.6.0
openai==1.3.7
anthropic==0.7.8

# Document Processing
python-docx==1.1.0
PyPDF2==3.0.1
python-pptx==0.6.23

# Email
emails==0.6.0

# File Storage
boto3==1.34.0
google-cloud-storage==2.10.0

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Utilities
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
typer==0.9.0

# Date & Time
python-dateutil==2.8.2
pytz==2023.3

# Data Processing
pandas==2.1.4
numpy==1.25.2

# Web Scraping
beautifulsoup4==4.12.2
selenium==4.15.2
playwright==1.40.0

# API Integration
linkedin-api==2.0.0
google-api-python-client==2.108.0
google-auth==2.25.2
google-auth-oauthlib==1.1.0

# Validation
email-validator==2.1.0
phone==0.1.0

# Serialization
orjson==3.9.10
msgpack==1.0.7

# Async Utilities
asyncio-mqtt==0.16.1
aiofiles==23.2.1

# Configuration
dynaconf==3.2.4

# Health Checks
healthcheck==1.3.0

# Rate Limiting
slowapi==0.1.9

# CORS
fastapi-cors==0.0.6 